Dropped support for Python 3.4, and added 3.6-3.8 to the test matrix
PyYAML no longer supports Python 3.4, and installation doesn't correctly
fall back to a version that works. We could probably find a way to
explicitly depend on an older version when installing under Python 3.4.
But this is a pretty clear signal that the ecosystem has moved on, and
so should Jasmine.